+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.12.0] - 2026-02-06
 ### Added
-- Upcoming changes...
+- Added `ComponentStatus` domain model with `StatusCode` type and constants (`ComponentNotFound`, `InvalidPurl`, `ComponentWithoutInfo`, `Success`, `InvalidSemver`)
+- Added `StatusCodeToErrorCode` function to map domain status codes to protobuf `ErrorCode` values
 
 ## [0.11.0] - 2026-01-30
 ### Changed
@@ -87,5 +90,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.7.0]: https://github.com/scanoss/go-grpc-helper/compare/v0.6.0...v0.7.0
 [0.8.0]: https://github.com/scanoss/go-grpc-helper/compare/v0.7.0...v0.8.0
 [0.9.0]: https://github.com/scanoss/go-grpc-helper/compare/v0.8.0...v0.9.0
-[0.11.0]: https://github.com/scanoss/go-grpc-helper/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/scanoss/go-grpc-helper/compare/v0.9.0...v0.10.0
+[0.11.0]: https://github.com/scanoss/go-grpc-helper/compare/v0.10.0...v0.11.0
+[0.12.0]: https://github.com/scanoss/go-grpc-helper/compare/v0.11.0...v0.12.0